Update backdrop.md
Line 12. It's unclear what the answer is intended to do. Is it discouraging use in Flutter? When it says you can achieve the result with "existing components and classes" does that mean it's recommending using something other than the backdrop component, or that you can hack the backdrop in a way that's ... something? Unclear what the 3rd sentence means, basically.

Line 16. When you say "a view" do you literally mean the Android definition of a view? https://developer.android.com/reference/android/view/View

Line 16. Does "reveal information behind it" mean reveal the back layer? Or something else? 

Line 19. "You will have a separate widget to build out the front layer" Is this "separate widget" referring to the stateful widget or something else? If something else, what is that else? 

Line 36. By "secondary" do you mean not dominant in the sense you need to take an action to access it?

Line 36. How is the back layer "persistent"? It's always there, but not always open (and I think often the spec uses persistent to mean active/open/inyourface.

Line 42. Does this mean a single front or back layer, as in not both? Or does it mean both, but is unique in some other way? Unclear on phrasing.

Line 42. "containing the backdrop to a vertical back and forth interaction" what does this mean?